Reorganized folder structure and tweaked pom.xml
