+++ pom.xml
@@ -1,6 +1,6 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>FreeHoldEm</groupId>
+  <groupId>edu.uab.cis.zbb</groupId>
   <artifactId>FreeHoldEm</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <dependencies>
@@ -8,10 +8,10 @@
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>4.11</version>
+      <scope>test</scope>
     </dependency>
   </dependencies>
-  <build>
-    <sourceDirectory>src</sourceDirectory>
+  <build> 
     <plugins>
       <plugin>
         <artifactId>maven-compiler-plugin</artifactId>
@@ -40,4 +40,9 @@
       </plugin>
     </plugins>
   </build>
+  <properties>
+  	<encoding>UTF-8</encoding>
+  	<maven.compiler.source>1.8</maven.compiler.source>
+  	<maven.compiler.target>1.8</maven.compiler.target>
+  </properties>
 </project>
