+++ src/main/java/edu/uab/cis/zbb/FreeHoldEm/Tuple.java
@@ -1,4 +1,4 @@
-package application;
+package edu.uab.cis.zbb.FreeHoldEm;
 public class Tuple {
 
     public int x;
